backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
pyyaml>=6.0